Include backing node request id in metastatic log
When we are satisfying a metastatic request by creating a new backing
node, we can end up logging the message:

  Assigned node 001 to backing node None slot 0

Which is not helpful when trying to find out which backing node 001
was assigned to.  We don't know the node id yet since it hasn't been
created, but in this case we can log the request id.  That can then
be associated with the actual node id in later log entries.

Change-Id: I409917319ab592c997f47e7f69636860fad8c2c9