Update patch set 1
Patch Set 1: Code-Review-1

According to the traceback, the error was that the diskimage specified in the provider did not exist as a top-level diskimage config object.  The validation to correct that should check for that, and doesn't need to be concerned with labels at all.

Also, it would be great if the test fixture were simpler.  The demonstration for this problem should be very small (only a few lines).

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1