Add config option to limit ephemeral storage on K8s Pod labels
This adds config options for limiting the amount of ephemeral storage
allocatable by a container of a K8s Pod-type label.
This optional config translates to K8s settings

* spec.containers[].resources.limits.ephemeral-storage
* spec.containers[].resources.requests.ephemeral-storage

This is to provide a mechanism that prevents Pods from filling up their
hosts storage and thereby interfering with or breaking other workloads
on the same host (esp. on shared clusters).

Like for cpu and memory limits, a pool-scoped default can also be
specified.

Change-Id: I23e90ae53cc2b2eb0e51cc9e3dc5802c86cc0ac9