WIP: AWS add support for multiple instance types
This allows operators to configure one or more fallback instance types.

When attempting to use highly specialized instance types in AWS, it is
not unlikely to encounter capacity constraints.  In cases where an operator
expects a job to be able to run on a less-specialized instance type,
Nodepool can now be configured to retry the launch wit the fallback
type.  In essence, operators can now indicate that for the purposes of
a given label, multiple AWS instance types are satisfactory.

Because the instance types may have distinct quota values, some updates
are needed.  First, we now calculate quota used by Nodepool nodes using
the resources attribute of the node rather than asking the driver what
the quota values are for the node's label (since a node's label may
no longer correspond 1:1 to an instance type).  Second, we ensure that
all quota-supporting drivers store resource information on the node to
support this change.

Change-Id: Ia041f32a481220b6a6fe3da90adc70e0cfb959d4