Use per-instance LRU caches on statemachine adapters
Several drivers used the lru_cache decorator which creates a global
cache rather than an instance-specific cache.  This could cause us
to leak memory over time as adapters are replaced (this happens when
the config for a provider changes and the driver is reloaded).

To correct this, create the cache in the constructor and wrap the
methods there.

The Azure driver wasn't subject to the memory leak because it had
a 24 hour TTL, but the intent of that method is the same as the others
(images don't change), so let's use the same approach for all the
drivers.

Change-Id: Ifb0f2706c9ff6daf9ad84213d5fa4a52e123c39f