Cleanup local builds without .d folder
When using a custom image build tool the .d folder does not exists and the
cleanup of those local builds is skipped.
For these reason we should not relei on the .d folder, instead we should look
on all create images files and determinate the local builds based on these.

Change-Id: I1c60af3347868089ebe489ddcadfbad2dc8fadde