Serve all paused handlers before unpausing
The launcher implementation assumed that only one request handler can be
paused at any given point in time. However, this is not true when e.g. the
request handler accepts multiple requests that all run into a quota
during launch.

The consequence of this is that the pool is unpaused too early and we
might accept other node requests until the provider is paused again.
This could lead to a starvation of earlier paused handlers as they were
fulfilled in a LIFO fashion.

To fix this edge case we will store a set of paused request handlers and
only unpause the provider when there are no paused handlers anymore.
Paused handlers are now also run in priority order.

Change-Id: Ia34e2844533ce9942d489838c4ce14a605d79287