Fix gce test pausing at quote
The existing test for gce machine boots assumed that after the first
boot was complete it would be at quote causing the second request to
pause. Unfortunately, we had quota for more than one instance which
meant that the test proceeded with fulfilling the second request. If
that was request was fulfilled quickly enough we would fail because the
status would change from PENDING to FULFILLED.

Fix this by updating our cloud quota to allow for a single instance
instead which will cause the pause to happen.

Change-Id: I4ebeb684a2c99ad1af12420f4d882777c38fb156