Add debug logs to the static driver
We've got a test race (that is probably a real bug) that is difficult to
understand because it implies that the static driver's start() method
may be called more often than expected. To understand when start() is
called and when the two methods that register nodes are called we add
logging to capture when these steps are taken and not just when they
fail.

Change-Id: I9fea269dc08eef8cecc0a2cf803d2dbe5e05fc0f