Stop provider managers before starting new ones
The way we reconfigure our driver provider managers when their configs
change is to discard the old managers and start new ones with the new
config. Previously, we started the new provider manager before stopping
the old ones. This is problematic because provider managers run periodic
tasks which means the old manager could run a periodic task as the new
manager is starting which interferes with the operation of the new
manager.

This is particularly true of the static driver as it must maintain
significant state locally in nodepool as there is no cloud to rely on
for this. What we observed was that if you updated configs to change the
number of slots or remove a node the old manager's cleanup routine could
run as the new manager is starting up. If this happens the old config
will be applied until some additional event occurs to reapply the new
managers config that has been overwritten by the rogue old manager.

To address this we simply stop managers before starting new ones.

Change-Id: I93519d0c6f4ddf8a417d837f6ae12a30a55870bb