Lock around config updates and provider manager usage
Nodepool associates provider managers with its running config. It then
uses that running config to lookup provider managers to run tasks
against a provider. Tasks like cleanup of resources and handler
assignment occur this way. Problem is that when we update our config we
may have both an old provider manager that is assocaited with the config
and a new provider manager that is starting up and performing
initialization tasks. These separate activities may conflict with one
another leading to problems.

We have noticed this with the static driver in particular. The reason
for this is that when the static driver is started it loads its config
and applies that state to zookeeper. But the old provider manager
cleanup tasks may run which overwrite that newly loaded config. This
will persist until some other event causes the driver to reconcile its
configs with zookeeper again.

To address this we lock around the reconfiguration config updates which
set our configs and where we retrieve provider managers out of the
config. This should ensure actions taken by the provider manager occur
while it is the canonical provider manager for a config state and not
the old but unremoved config.

Change-Id: I93519d0c6f4ddf8a417d837f6ae12a30a55870bb