Unpin openstacksdk
This unpins openstacksdk and has nodepool use the latest version. We do
this becaues 0.61.0 and older are not compatible with python 3.11.
0.99.0 and newer should be, but they may introduce compatibility issues
due to major rewrites in the SDK.

We know 0.99.0 through 0.102.0 do not work due improper handling of
nova api microversions when setting network information during server
creation. 0.103 fixes this and I've tested the fix with the proposed
openstack driver rewrite to use statemachine with openstacksdk 0.103.0
and it appears to function against rax and iweb. While it is possible
there are other issues say in image uplodaing the impact htere is lower
and we can always revert if necessary.

Change-Id: Iace27698a91aa2e4b5f169f8a2c1613e5dd37356