Fix AWS instance_profile tests for moto v4
Moto v4 has added support for iam instance profiles, which means
our tests are now failing since we have two tests that request
an instance be created with a profile association, but the profile
did not exist in moto due to the lack of support.

Now that support has been added, we create a dummy instance profile
and use that for the association.  We can now also validate that
the association was made as expected.

Change-Id: I2839b7e1536a6305e44678dc9b22c71f3eb55677