Fix openstack image deletion with newer sdk
Openstacksdk version 0.103.0 removed an informal API we were using
which accepted an abbreviated dictionary as input to the delete_image
method.

It now requires either a complete image object or just a name_or_id,
so we now pass in the id.

The sdk min version is increased since older versions have not been
tested with this.

Change-Id: I7df276ab76e9b8fc17612853b474fec414dae977