Dockerfile: use containernetworking-plugins
It seems the containerfile elements are currently failing to have any
networking, meaning they can't update packages, etc.

There's some podman warnings along the lines of

 Error validating CNI config file
 /etc/cni/net.d/87-podman-bridge.conflist: [failed to find plugin
 \"bridge\" ...

which suggests to me containernetworking-plugins isn't installed.
Honestly I'm not sure why we aren't installing it ... it has iptables
as a dependency so we don't need to explicilty install that.

Add it to the install here

Change-Id: Ie1fe48691f44aa847859cf294404dd9e8b03cce8