Fix AWS quota limits for vCPUs
In the AWS adapter, when getting the quota for an instance type, set
the quota for the AWS service quota code to be the number of vCPUs
rather than the number of cores. The number of vCPUs is typically
twice the number of cores. This fixes "VcpuLimitExceeded" errors from
AWS.

Change-Id: I880e6abb84b0527363893576057aa105a5a448a5