Dockerfile: remove unstable usage
We are having problems with the ARM64 nodepool-builder container that
appears to be because unstable now has a Python 3.11 package, and
pulling it in conflicts with the container-built version in
/usr/local.  It appears the /usr/local install isn't fully isolated
from the system install <insert reason here, it's not clear>.

We can hopefully avoid this by dropping the unstable dependency, which
is not great to have anyway.

stable-bpo has debootstrap 1.0.128 which is the same as unstable.  So
we can switch that to install from backports.

The issues with the clone3() system call are included in podman 3.0.1
which is in the stable repos now [1].  So we drop all the unstable
pulling of podman related packages.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=995777

Change-Id: Ib3dfda3df69e7ab359b96cd1865e47c7e1e8047b