driver/openstack: order flavor results
I hit a situation in a new cloud where I had defined two flavors with
the same amount of RAM "opendev-control" and "opendev".  The config
had min-ram set [1] with a flavor-name of "opendev" -- I expected it
to match the exact name first, but nodepool was choosing
"opendev-control".

I guess the default order returned by the cloud is flavorid [2].  I'd
propose that sorting on a tuple of (ram, name) -- so that we match
names in alphabetical order -- is a more intuitive way to run the
match.

Documentation is updated, and a release note added.

[1] which actually we didn't want, really, because we wanted to
    exactly match the flavor:
     https://review.opendev.org/c/openstack/project-config/+/870677
[2] https://docs.openstack.org/api-ref/compute/?expanded=list-flavors-detail#list-flavors

Change-Id: I268dd598ca9f1b617c5062b41ad27d0305df60b9