Introduce driver for openshift virtualmachines
With openshift virtualization it makes sense to integrate this resource
type with nodepool. This change proposes a new driver to create and delete
virtualmachines dynamically using a VirtualMachine manifest template
along with a service exposing an ssh endpoint. The template file needs
to be mounted to the deployment and pointed to in the nodepool config
using the openshiftlabels.

Both the virtualmachine and service will be created on launch and
deleted during cleanup.

Change-Id: Ibd8a24edf7d35c079363b2110a65350029abe3ad