Speed up openstack server listing
The old openstack driver used the "bare=True" argument to list_servers
and then relied on wait_for_server to expand the network information
on the server when it was active.  The new driver omits the bare
argument meaning that neutron is consulted for all server lists.

This can be very slow on large systems, so let's switch back to
bare=True, but instead of calling wait_for_server (which we are
unable to do in the state machine framework since state machine
runs are expected to be quick), we add the interface data ourselves
when refreshing information about active servers.

Change-Id: Ia7dce406e14b06166d5fa1d5d9bea0e661122dde