Revert "Update git"
This reverts commit 938d28abe820cb9164c2f9765386c2354ecc374d.

Upstream Debian has updated git to 2.30.2-1+deb11u1 which patches git
for the issues we manually patched to cover. We don't need the manual
patch anymore and can switch to the distro hence this revert.

Reviewers should double check that the image build process installs the
expected 2.30.2-1+deb11u1 version.

Change-Id: I0240b6a311dff7240571b70b814223ec5761fcc3