Offload openstack delete api calls to an executor
Use a thread pool exeuctor to run openstack API server delete
calls since they can take an appreciable amount of time to
return.

This also increases the number of worker threads to 16 to account
for the extra executor usage.

It also sets the number of keyscan workers to 16 (which is probably
an increase; it was undefined before).

This allows for more parallel operation at the cost of a little more
threading overhead.

Change-Id: I55984b21677eb1291d11a0e80945015f68f67b5e