Fix race in metastatic allocation
Metastatic backing nodes are deleted once they are idle for a
period of time.  That is checked periodically in a cleanup thread.

If a new request arrives while the cleanup thread is running,
the new request may be assigned to a backing node that is in
the process of being deleted.

To address this, we wrap all of the accesses to the
backing_store_records attribute in a mutex.  The typical case has
us only accessing this in a single thread: the state machine runner,
so there should be no additional delay, unless the cleanup thread
is running.

The race is very tight and not able to be tested with the current
test framework, so no test is added.

Change-Id: I8d2f0ba03482693df5fd7810e580c8b1a05722e8