Add a stop thread to statemachine
Providers have a "stop" and a "join" method so that when new
configuration arrives, the launcher can tell all existing
providers to quickly "stop".  If the program is exiting, they
can optionally be "joined".

In the statemachine "stop" method, we wait for all statemachines
to be complete before shutting down the keyscan threadpool,
because otherwise shutting down the threadpool would cause any
existing statemachines to fail at the end of the process.

However, this causes the "stop" method to take a long time, and
during that time, new requests can be assigned to the provider,
extending the process even more.

To allow the stop method to quickly return, yet also support
the extended sequence needed due to the keyscan threadpool, this
change causes the stop method to spawn a thread dedicated to
stopping the provider manager.  The join method now waits for
that thread to complete.  This restores the statemachine
provider manager to the intended "fire and forget" approach
of the stop method.

Change-Id: I430c70ce0daa7fc28cbd43ecc64c7a974239950d