Allow to pass ImagePullSecrets to Kubernetes pods
This is currently only supported for the Openshift driver but it's quite useful (and needed)
for the Kubernetes driver as well. The implementation follows the same strategy as for the
OpenShift driver by allowing to set the  key to the Pod labels.

One caveat is that this exects the secret to exist in the namespace, therefore, some external
mechanism needs to take care of creating such secret there.

Change-Id: I976290e72cd59c335f6bef22e364a4d5c0a7b554