Process paused requests first
We observed a starvation problem in the following scenario leading to
gaps in request processing of several minutes.

Scenario:
- there are many pending node request
- a request handler gets paused by running into quota

In this case nodepool loops over all node requests while deferring all
requests due to the paused handler until it retries the paused
handler. In our case this took 10 minutes until it unpaused and
continued normal work normally. As long as the request queue is long
this starts over as soon as it reaches the cloud quota again.

The paused handlers are processed before looping through the node
requests. Similar to fixing the starvation problem of removing
completed handlers we process them within the loop as well.

Change-Id: Iadacd4969c883574d8947e8ab2313e42820cb298